requirements.txt
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
flask_admin